Crate cplfs_sol

Root file of your solution

This file

There are two types of projects or crates in Rust: libraries and executables. The difference should be obvious. The root of a library project is src/lib.rs (like this file) and the root of an executable project is src/main.rs.

The root file contains some important things:

It provides some documentation for the crate. You are reading it right now.

It contains annotations that are true for the whole crate. See the source code of this file for an example.

It declares dependencies on external crates. Again, see the source code.

It declares the modules of which the crate consists using the mod keyword. Don't confuse this with the use keyword. If you add any modules to this project, you have to declare at the end of this file using the mod keyword. Unless they are submodules of another module, in which case you have to declarate at the top of that module. In case you want them to be public (and visible in the documentation too), use pub mod instead of mod.

See the chapter on Crates and Modules in the Rust book for more information.

The source code of this file contains more information.

If you are looking at the documentation of this file, the list on the left side contains all the crates that this crate transitively depends on.

The next thing to look at is the a_block_support module.

Rust Version

TODO: indicate which stable version of Rust you are using. If you are using 1.47, you don't have to do anything. Otherwise, replace the version below with the output of rustc --version.

VERSION: rustc 1.47 (2020-10-08)

Modules

a_block_support
File system with block support

b_inode_support
File system with inode support

c_dirs_support
File system with directory support

d_path_support
File system with path support

e_inode_RW_support
File system with inode support + read and write operations on inodes

f_indirect_inodes
File system with support for inodes that have indirect blocks too. Reimplementation of the inodes from the base project. This additional assignment requires completion of assignment e, so go and do that one first if you haven't yet.

g_caching_inodes
File system with support for inode caching
Reimplementation of the inodes from the base project.