Module cplfs_sol::b_inode_support

File system with inode support

Create a filesystem that has a notion of inodes and blocks, by implementing the FileSysSupport, the BlockSupport and the InodeSupport traits together (again, all earlier traits are supertraits of the later ones).

Make sure this file does not contain any unaddressed TODOs anymore when you hand it in.

Status

TODO: Replace the question mark below with YES, NO, or PARTIAL to indicate the status of this assignment. If you want to tell something about this assignment to the grader, e.g., you have a bug you can't fix, or you want to explain your approach, write it down after the comments section. If you had no major issues and everything works, there is no need to write any comments.

COMPLETED: ?

COMMENTS:

...

Structs

InodeFS
Main struct file for the Inode File System

Enums

InodeFSError
Main error file for Inode File system

Type Definitions

FSName
You are free to choose the name for your file system. As we will use automated tests when grading your assignment, indicate here the name of your file system data type so we can just use FSName instead of having to manually figure out the name. TODO: replace the below type by the type of your file system