Type Definition cplfs_sol::b_inode_support::FSName

type FSName = InodeFS;

You are free to choose the name for your file system. As we will use automated tests when grading your assignment, indicate here the name of your file system data type so we can just use FSName instead of having to manually figure out the name. TODO: replace the below type by the type of your file system